Implement controllable R and U rotations (messy)
