+++ src/lib/components/2d/Net.svelte
