+++ src/lib/components/2d/Sticker.svelte
